requirements.txt
numpy==1.26.4
torch==2.2.1
tqdm==4.66.2
scikit-learn==1.4.1
segyio==1.9.12
pillow==10.2.0